Started spliting up to smaller classes
