+++ Runtime/Scripts/Core/LogHelpMethods.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f6362e152bcb65046a4ac91d46a6af1f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
