+++ Runtime/Scripts/Core/SytemInfoLogger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: defd51dd62a3c9a439b44a3e2218df2b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
